fix(KEN-877): Address review - rebase onto main [no-changelog]
Two things the rebase onto origin/main needed, neither a behaviour change.

Main dropped the fail_after parameter from apply::execute in 9a9ac27c7.
Three call sites this branch added after the two conflicted commits still
passed None, in settings_edits.rs and the settings_seed/scope.rs helpers
that replaced settings_seed_apply.rs. Main never saw the split files, so
its own sweep could not reach them. A tree-wide sweep now finds no
three-argument apply::execute left.

expansion.rs then crossed the 400-line cap at 406. Main added 10 lines to
it and this branch added 21, and each side was under the cap alone. The
seam size-ratchet asks for was already there: skills_installed is read by
ops::add and by nothing else, and answers what a manifest has installed
without planning anything, so it moves to engine/installed.rs with its own
doc. That also puts back the doc comment for expand, which skills_installed
had been inserted in front of and taken over, leaving expand undocumented
and skills_installed carrying a paragraph about a held parameter it does
not take.
